Task successfully connected to UI! All difficulty levels are working

+++ engine/src/engine/Engine.java
@@ -63,7 +63,7 @@ public interface Engine extends Serializable {
 
     void setDMParamsFromUI(int agentCountChosen, String difficulty, int missionSize);
 
-    Integer calculateTotalMissionsAmount();
+    Double calculateTotalMissionsAmount();
 
     String processWord(String word);
